Moved Container to DataModel + Updated project target framework
