+++ Project/ModelLabsApp/App.config
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
+<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <runtime>
     <loadFromRemoteSources enabled="true"/>
@@ -6,10 +6,7 @@
   <system.serviceModel>
     <client>
       <!-- Network Model Service-->
-      <endpoint address="net.tcp://localhost:10000/NetworkModelService/GDA/"
-                behaviorConfiguration="NMSClientBehavior"
-                binding="netTcpBinding" bindingConfiguration="netTcpBindingNoSecurity"
-                contract="FTN.ServiceContracts.INetworkModelGDAContract" name="NetworkModelGDAEndpoint">
+      <endpoint address="net.tcp://localhost:10000/NetworkModelService/GDA/" behaviorConfiguration="NMSClientBehavior" binding="netTcpBinding" bindingConfiguration="netTcpBindingNoSecurity" contract="FTN.ServiceContracts.INetworkModelGDAContract" name="NetworkModelGDAEndpoint">
         <identity>
           <dns value="localhost"/>
         </identity>
@@ -18,12 +15,10 @@
 
     <bindings>
       <netTcpBinding>
-        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00"
-				 openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647"
-				  maxReceivedMessageSize="2147483647">
+        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00" openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647" maxReceivedMessageSize="2147483647">
           <readerQuotas maxStringContentLength="2147483647" maxArrayLength="2147483647"/>
           <security mode="None">
-            <transport clientCredentialType="None" protectionLevel="None" />
+            <transport clientCredentialType="None" protectionLevel="None"/>
           </security>
         </binding>
       </netTcpBinding>
@@ -41,4 +36,4 @@
     </behaviors>
 
   </system.serviceModel>
-</configuration>
+<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/></startup></configuration>
