+++ Project/NetworkModelService/ServiceHost/NetworkModelServiceSelfHost/app.config
@@ -1,9 +1,9 @@
-<?xml version="1.0" encoding="utf-8" ?>
+<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <system.diagnostics>
     <trace autoflush="true">
       <listeners>
-        <add type="System.Diagnostics.TextWriterTraceListener" name="TextWriter" initializeData="../NetworkModelService.log" />
+        <add type="System.Diagnostics.TextWriterTraceListener" name="TextWriter" initializeData="../NetworkModelService.log"/>
       </listeners>
     </trace>
     <switches>
@@ -12,30 +12,29 @@
            2 = Warning - Gives errors and warnings
            3 = Info    - Gives more detailed error information
            4 = Verbose - Gives verbose trace information.    -->
-      <add name="TraceLevel" value = "Info" />
+      <add name="TraceLevel" value="Info"/>
     </switches>
   </system.diagnostics>
 
   <connectionStrings>
-    <add name="networkModelconnectionString"	connectionString="../NetworkModelData.data"/>
+    <add name="networkModelconnectionString" connectionString="../NetworkModelData.data"/>
   </connectionStrings>
   <system.web>
-    <compilation debug="true" />
+    <compilation debug="true"/>
   </system.web>
   <!-- When deploying the service library project, the content of the config file must be added to the host's 
   app.config file. System.Configuration does not support config files for libraries. -->
   <system.serviceModel>
     <services>
       <service behaviorConfiguration="NMS.ServiceBehavior" name="FTN.Services.NetworkModelService.GenericDataAccess">
-        <endpoint address="" binding="netTcpBinding" contract="FTN.ServiceContracts.INetworkModelGDAContract" 
-                  bindingConfiguration="netTcpBindingNoSecurity">
+        <endpoint address="" binding="netTcpBinding" contract="FTN.ServiceContracts.INetworkModelGDAContract" bindingConfiguration="netTcpBindingNoSecurity">
           <identity>
-            <dns value="localhost" />
+            <dns value="localhost"/>
           </identity>
         </endpoint>
         <host>
           <baseAddresses>
-            <add baseAddress="net.tcp://localhost:10000/NetworkModelService/GDA/" />
+            <add baseAddress="net.tcp://localhost:10000/NetworkModelService/GDA/"/>
           </baseAddresses>
         </host>
       </service>
@@ -43,12 +42,10 @@
 
     <bindings>
       <netTcpBinding>
-        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00"
-				 openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647"
-				  maxReceivedMessageSize="2147483647">
+        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00" openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647" maxReceivedMessageSize="2147483647">
           <readerQuotas maxStringContentLength="2147483647" maxArrayLength="2147483647"/>
           <security mode="None">
-            <transport clientCredentialType="None" protectionLevel="None" />
+            <transport clientCredentialType="None" protectionLevel="None"/>
           </security>
         </binding>
       </netTcpBinding>
@@ -57,16 +54,12 @@
     <behaviors>
       <serviceBehaviors>
         <behavior name="NMS.ServiceBehavior">
-          <serviceMetadata httpGetEnabled="false" />
-          <serviceThrottling
-           maxConcurrentCalls="1000"
-           maxConcurrentSessions="1000"
-           maxConcurrentInstances="1000"
-          />
-          <serviceDebug includeExceptionDetailInFaults="true" />
+          <serviceMetadata httpGetEnabled="false"/>
+          <serviceThrottling maxConcurrentCalls="1000" maxConcurrentSessions="1000" maxConcurrentInstances="1000"/>
+          <serviceDebug includeExceptionDetailInFaults="true"/>
         </behavior>
       </serviceBehaviors>
     </behaviors>
 
   </system.serviceModel>
-</configuration>
+<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/></startup></configuration>
