+++ Project/NetworkModelService/Test/NMSTestClient/app.config
@@ -7,7 +7,7 @@
   <system.diagnostics>
     <trace autoflush="true">
       <listeners>
-        <add type="System.Diagnostics.TextWriterTraceListener" name="TextWriter" initializeData="../TestClient.log" />
+        <add type="System.Diagnostics.TextWriterTraceListener" name="TextWriter" initializeData="../TestClient.log"/>
       </listeners>
     </trace>
     <switches>
@@ -16,17 +16,14 @@
            2 = Warning - Gives errors and warnings
            3 = Info    - Gives more detailed error information
            4 = Verbose - Gives verbose trace information.    -->
-      <add name="TraceLevel" value = "Info" />
+      <add name="TraceLevel" value="Info"/>
     </switches>
   </system.diagnostics>
 
 	<system.serviceModel>
 		<client>
 			<!-- Network Model Service-->
-			<endpoint address="net.tcp://localhost:10000/NetworkModelService/GDA/" 
-                behaviorConfiguration="NMSClientBehavior"
-                binding="netTcpBinding" bindingConfiguration="netTcpBindingNoSecurity"
-                contract="FTN.ServiceContracts.INetworkModelGDAContract" name="NetworkModelGDAEndpoint">
+			<endpoint address="net.tcp://localhost:10000/NetworkModelService/GDA/" behaviorConfiguration="NMSClientBehavior" binding="netTcpBinding" bindingConfiguration="netTcpBindingNoSecurity" contract="FTN.ServiceContracts.INetworkModelGDAContract" name="NetworkModelGDAEndpoint">
 				<identity>
 					<dns value="localhost"/>
 				</identity>
@@ -35,12 +32,10 @@
 
     <bindings>
       <netTcpBinding>
-        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00"
-				 openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647"
-				  maxReceivedMessageSize="2147483647">
+        <binding name="netTcpBindingNoSecurity" closeTimeout="00:30:00" openTimeout="00:30:00" sendTimeout="00:10:00" receiveTimeout="10:30:00" maxBufferSize="2147483647" maxReceivedMessageSize="2147483647">
           <readerQuotas maxStringContentLength="2147483647" maxArrayLength="2147483647"/>
           <security mode="None">
-            <transport clientCredentialType="None" protectionLevel="None" />
+            <transport clientCredentialType="None" protectionLevel="None"/>
           </security>
         </binding>
       </netTcpBinding>
@@ -58,4 +53,4 @@
     </behaviors>
     
 	</system.serviceModel>
-</configuration>
+<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/></startup></configuration>
